# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {},
        "license": {
            "name": "Apache 2.0"
        }
    },
    "paths": {
...
        "/api/v1/base/organizations": {
            "get": {
                "description": "查询组织列表",
                "tags": [
                    "Organization"
                ],
                "summary": "查询组织列表",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "是否仅查询count",
                        "name": "-",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "当前页",
                        "name": "current",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "是否查询所有数据",
                        "name": "find_all",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "组织ID集合(逗号分隔)",
                        "name": "ids",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页大小",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "parent_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "模糊搜索(搜索 组织名称/备注)",
                        "name": "queryValue",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "角色ID",
                        "name": "role_id",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "是否显示组织树(获取组织树、职位列表.不包含禁用的组织和职位)",
                        "name": "show_details",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "是否显示组织的职位列表",
                        "name": "show_positions",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "状态(1:启用 2:禁用)",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ginx.ListResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ginx.OperateResult"
                        }
                    }
                }
            },
            "post": {
                "description": "创建组织信息(职位没有ID且职位的组织ID为0，则创建职位)",
                "tags": [
                    "Organization"
                ],
                "summary": "创建组织",
                "parameters": [
                    {
                        "description": "创建组织参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/schema.Organization"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.IDResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ginx.OperateResult"
                        }
                    }
                }
            }
        },
        "/api/v1/base/organizations/:id": {
            "delete": {
                "description": "删除组织和组织的职位",
                "tags": [
                    "Organization"
                ],
                "summary": "删除组织",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "组织ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ginx.OperateResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ginx.OperateResult"
                        }
                    }
                }
            }
        },
        "/api/v1/base/organizations/create-notifications": {
            "get": {
                "description": "获取组织树(不包含职位和禁用的组织)",
                "tags": [
                    "Organization"
                ],
                "summary": "获取创建通知公告的组织树",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/schema.Organization"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ginx.OperateResult"
                        }
                    }
                }
            }
        },
        "/api/v1/base/organizations/create-user": {
            "get": {
                "description": "获取组织树、职位列表(不包含禁用的组织和职位)",
                "tags": [
                    "Organization"
                ],
                "summary": "获取创建用户的组织树",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/schema.Organization"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ginx.OperateResult"
                        }
                    }
                }
            }
        },
        "/api/v1/base/organizations/edit": {
            "get": {
                "description": "获取组织树、职位列表(包含禁用的组织、职位)",
                "tags": [
                    "Organization"
                ],
                "summary": "获取可编辑的组织树",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/schema.Organization"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ginx.OperateResult"
                        }
                    }
                }
            }
        },
        "/api/v1/base/organizations/include-user": {
            "get": {
                "description": "获取包含用户的组织树、职位列表(不包含禁用的组织和职位)",
                "tags": [
                    "Organization"
                ],
                "summary": "获取人员组织架构树",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/schema.OrganizationTree"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ginx.OperateResult"
                        }
                    }
                }
            }
        },
        "/api/v1/base/organizations/update-basic-info/:id": {
            "put": {
                "description": "更新组织的基础信息",
                "tags": [
                    "Organization"
                ],
                "summary": "更新组织的基础信息",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "组织ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "组织基础信息参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schema.Organization"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ginx.OperateResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ginx.OperateResult"
                        }
                    }
                }
            }
        },
        "/api/v1/base/organizations/user-join/:id": {
            "put": {
                "description": "用户加入指定组织",
                "tags": [
                    "Organization"
                ],
                "summary": "用户加入指定组织",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "组织ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "用户ID参数集合",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ginx.OperateResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ginx.OperateResult"
                        }
                    }
                }
            }
        },
        "/api/v1/base/organizations/user-remove/:id": {
            "put": {
                "description": "用户移出指定组织",
                "tags": [
                    "Organization"
                ],
                "summary": "用户移出指定组织",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "组织ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "用户D参数集合",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ginx.OperateResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ginx.OperateResult"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "common.IDResult": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "0"
                }
            }
        },
        "common.PaginationResult": {
            "type": "object",
            "properties": {
                "current": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "ginx.ListResult": {
            "type": "object",
            "properties": {
                "list": {},
                "pagination": {
                    "$ref": "#/definitions/common.PaginationResult"
                }
            }
        },
        "ginx.OperateResult": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "错误码",
                    "type": "integer"
                },
                "message": {
                    "description": "错误信息",
                    "type": "string"
                }
            }
        },
...
        "schema.Organization": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "created_at": {
                    "description": "创建时间",
                    "type": "string"
                },
                "creator": {
                    "description": "创建者",
                    "type": "integer"
                },
                "id": {
                    "description": "唯一标识",
                    "type": "integer"
                },
                "memo": {
                    "description": "备注",
                    "type": "string"
                },
                "name": {
                    "description": "组织名称",
                    "type": "string"
                },
                "parent_id": {
                    "description": "父级组织ID",
                    "type": "integer"
                },
                "positions": {
                    "description": "组织的职位列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.Position"
                    }
                },
                "role_id": {
                    "description": "组织的基础角色",
                    "type": "integer"
                },
                "sequence": {
                    "description": "排序值",
                    "type": "integer"
                },
                "son_organizations": {
                    "description": "下属组织列表",
                    "allOf": [
                        {
                            "$ref": "#/definitions/schema.Organizations"
                        }
                    ]
                },
                "status": {
                    "description": "状态(1:启用 2:禁用)",
                    "type": "integer"
                },
                "updated_at": {
                    "description": "更新时间",
                    "type": "string"
                }
            }
        },
        "schema.OrganizationTree": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "created_at": {
                    "description": "创建时间",
                    "type": "string"
                },
                "creator": {
                    "description": "创建者",
                    "type": "integer"
                },
                "id": {
                    "description": "唯一标识",
                    "type": "integer"
                },
                "memo": {
                    "description": "备注",
                    "type": "string"
                },
                "name": {
                    "description": "组织名称",
                    "type": "string"
                },
                "parent_id": {
                    "description": "父级组织ID",
                    "type": "integer"
                },
                "positions": {
                    "description": "组织的职位列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.Position"
                    }
                },
                "role_id": {
                    "description": "组织的基础角色",
                    "type": "integer"
                },
                "sequence": {
                    "description": "排序值",
                    "type": "integer"
                },
                "son_organization_trees": {
                    "description": "下属组织列表",
                    "allOf": [
                        {
                            "$ref": "#/definitions/schema.OrganizationTrees"
                        }
                    ]
                },
                "status": {
                    "description": "状态(1:启用 2:禁用)",
                    "type": "integer"
                },
                "updated_at": {
                    "description": "更新时间",
                    "type": "string"
                },
                "users": {
                    "description": "组织用户列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.User"
                    }
                }
            }
        },
        "schema.OrganizationTrees": {
            "type": "array",
            "items": {
                "$ref": "#/definitions/schema.OrganizationTree"
            }
        },
        "schema.Organizations": {
            "type": "array",
            "items": {
                "$ref": "#/definitions/schema.Organization"
            }
        },
...
        "schema.Position": {
            "type": "object",
            "required": [
                "name",
                "organization_id"
            ],
            "properties": {
                "created_at": {
                    "description": "创建时间",
                    "type": "string"
                },
                "creator": {
                    "description": "创建者",
                    "type": "integer"
                },
                "id": {
                    "description": "唯一标识",
                    "type": "integer"
                },
                "memo": {
                    "description": "备注",
                    "type": "string"
                },
                "name": {
                    "description": "职位名称",
                    "type": "string"
                },
                "organization_id": {
                    "description": "职位的组织ID",
                    "type": "integer"
                },
                "role_id": {
                    "description": "职位的角色",
                    "type": "integer"
                },
                "sequence": {
                    "description": "排序值",
                    "type": "integer"
                },
                "status": {
                    "description": "状态(1:启用 2:禁用)",
                    "type": "integer"
                },
                "updated_at": {
                    "description": "更新时间",
                    "type": "string"
                }
            }
        },
...
        "schema.Role": {
            "type": "object",
            "required": [
                "name",
                "type"
            ],
            "properties": {
                "auditor_types": {
                    "description": "审核类型(逗号分隔)",
                    "type": "string"
                },
                "created_at": {
                    "description": "创建时间",
                    "type": "string"
                },
                "creator": {
                    "description": "创建者",
                    "type": "integer"
                },
                "id": {
                    "description": "唯一标识",
                    "type": "integer"
                },
                "memo": {
                    "description": "备注",
                    "type": "string"
                },
                "name": {
                    "description": "角色名称",
                    "type": "string"
                },
                "sequence": {
                    "description": "排序值",
                    "type": "integer"
                },
                "status": {
                    "description": "状态(1:启用 2:禁用)",
                    "type": "integer"
                },
                "type": {
                    "description": "角色类型(1.用户角色 2.组织角色 3.职位角色 4.用户组角色)",
                    "type": "integer"
                },
                "updated_at": {
                    "description": "更新时间",
                    "type": "string"
                }
            }
        },
...
        "schema.User": {
            "type": "object",
            "required": [
                "user_name"
            ],
            "properties": {
                "created_at": {
                    "description": "创建时间",
                    "type": "string"
                },
                "creator": {
                    "description": "创建者",
                    "type": "integer"
                },
                "extend_info": {
                    "description": "用户扩展信息",
                    "allOf": [
                        {
                            "$ref": "#/definitions/schema.UserExtendInfo"
                        }
                    ]
                },
                "id": {
                    "description": "唯一标识",
                    "type": "integer"
                },
                "organizations": {
                    "description": "用户所属的组织结构",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.Organization"
                    }
                },
                "password": {
                    "description": "用户密码",
                    "type": "string"
                },
                "positions": {
                    "description": "用户所属的职位",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.Position"
                    }
                },
                "roles": {
                    "description": "用户所属的角色(直接绑定给用户的角色)",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.Role"
                    }
                },
                "sequence": {
                    "description": "排序值",
                    "type": "integer"
                },
                "status": {
                    "description": "状态(1:启用 2:禁用)",
                    "type": "integer"
                },
                "updated_at": {
                    "description": "更新时间",
                    "type": "string"
                },
                "user_groups": {
                    "description": "用户所属的的用户组",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.UserGroup"
                    }
                },
                "user_name": {
                    "description": "用户名称",
                    "type": "string"
                }
            }
        },
        "schema.UserExtendInfo": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "创建时间",
                    "type": "string"
                },
                "creator": {
                    "description": "创建者",
                    "type": "integer"
                },
                "email": {
                    "description": "邮箱账号",
                    "type": "string"
                },
                "id": {
                    "description": "唯一标识",
                    "type": "integer"
                },
                "mobile_phone": {
                    "description": "移动手机",
                    "type": "integer"
                },
                "qq_account": {
                    "description": "QQ账号",
                    "type": "integer"
                },
                "real_name": {
                    "description": "用户昵称",
                    "type": "string"
                },
                "updated_at": {
                    "description": "更新时间",
                    "type": "string"
                },
                "user_id": {
                    "description": "用户ID",
                    "type": "integer"
                }
            }
        },
        "schema.UserGroup": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "created_at": {
                    "description": "创建时间",
                    "type": "string"
                },
                "creator": {
                    "description": "创建者",
                    "type": "integer"
                },
                "id": {
                    "description": "唯一标识",
                    "type": "integer"
                },
                "memo": {
                    "description": "备注",
                    "type": "string"
                },
                "name": {
                    "description": "用户组名称",
                    "type": "string"
                },
                "role_id": {
                    "description": "用户组的基础角色",
                    "type": "integer"
                },
                "sequence": {
                    "description": "排序值",
                    "type": "integer"
                },
                "status": {
                    "description": "状态(1:启用 2:禁用)",
                    "type": "integer"
                },
                "updated_at": {
                    "description": "更新时间",
                    "type": "string"
                }
            }
        }
    }
}